+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 [![CalVer](https://img.shields.io/badge/calver-YY.MINOR.MICRO-blue?style=flat-square)](https://calver.org/)
 
-## Dessine-moi 22.1.0 (unreleased)
+## Dessine-moi 22.1.0 (2022-07-19)
 
 ### Features
 
-- `Factory`: Support string-based lazy registration ({ghpr}`4`).
+- `Factory`: Support string-based lazy registration ({ghpr}`2`).
 
 ### Developer side